move use-page-visibility to react-hooks
