+++ libs/react-hooks/src/index.ts
@@ -1,13 +1,14 @@
-export * from './use-login-redirect'
-export * from './use-map-interaction'
 export * from './use-dataviews-layers'
-export * from './use-local-storage'
 export * from './use-debounce'
 export * from './use-layer-composer'
+export * from './use-local-storage'
+export * from './use-login-redirect'
 export * from './use-login'
+export * from './use-map-interaction'
 export * from './use-map-legend'
 export * from './use-memo-compare'
 export * from './use-navigator-online'
+export * from './use-page-visibility'
+export * from './use-previous'
 export * from './use-small-screen'
 export * from './use-tiles-state'
-export * from './use-previous'
