+++ libs/react-hooks/src/use-page-visibility/use-page-visibility.ts
@@ -1,0 +1,26 @@
+import { useCallback, useEffect, useState } from 'react'
+
+export function usePageVisibility() {
+  const [isVisible, setIsVisible] = useState(document.visibilityState === 'visible')
+  const [firstTimeVisible, setFirstTimeVisible] = useState(false)
+  const onVisibilityChange = useCallback(
+    () => setIsVisible(document.visibilityState === 'visible'),
+    []
+  )
+
+  useEffect(() => {
+    if (isVisible && !firstTimeVisible) {
+      setFirstTimeVisible(true)
+    }
+  }, [isVisible, firstTimeVisible])
+
+  useEffect(() => {
+    document.addEventListener('visibilitychange', onVisibilityChange, false)
+    return () => {
+      document.removeEventListener('visibilitychange', onVisibilityChange)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
+  return { isVisible, firstTimeVisible }
+}
